% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_eda}
\alias{calculate_eda}
\title{Calcuate the expected number of distinct alleles under a given multiplicity of infection (n)}
\usage{
calculate_eda(allele_freqs, n)
}
\arguments{
\item{allele_freqs}{Simplex of allele frequencies}

\item{n}{multiplicity of infection}
}
\description{
Calcuate the expected number of distinct alleles under a given multiplicity of infection (n)
}
