% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_he}
\alias{calculate_he}
\title{Calculate the expected heterozygosity from allele frequencies}
\usage{
calculate_he(allele_freqs)
}
\arguments{
\item{allele_freqs}{Simplex of allele frequencies}
}
\value{
Numeric scalar, the expected heterozygosity \code{1 - sum(p^2)}.
}
\description{
Calculate the expected heterozygosity from allele frequencies
}
\examples{
calculate_he(c(0.5, 0.3, 0.2))
}
