% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_map_allele_frequencies}
\alias{calculate_map_allele_frequencies}
\title{Calculate the maximum a posteriori (MAP) estimate of allele frequencies
from the estimated posterior joint distribution of allele frequencies}
\usage{
calculate_map_allele_frequencies(mcmc_results, merge_chains = TRUE)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\description{
Calculate the maximum a posteriori (MAP) estimate of allele frequencies
from the estimated posterior joint distribution of allele frequencies
}
\details{
Returns the sampled allele frequencies at the maximum sampled
likelihood
}
