% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_med_allele_freqs}
\alias{calculate_med_allele_freqs}
\title{Calculate the geometric median of the posterior distribution of allele
frequencies}
\usage{
calculate_med_allele_freqs(mcmc_results, merge_chains = TRUE)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
If \code{merge_chains} is \code{TRUE}, a named list with one numeric vector of
allele frequencies per locus. Otherwise a list with one such list per chain.
}
\description{
Calculate the geometric median of the posterior distribution of allele
frequencies
}
\details{
Returns the geometric median of the posterior distribution, defined
as the point minimizing the L2 distance from each sampled point.
}
\examples{
med_freqs <- calculate_med_allele_freqs(mcmc_results)
med_freqs[[1]]
}
