% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_naive_coi_offset}
\alias{calculate_naive_coi_offset}
\title{Calculate naive COI offset}
\usage{
calculate_naive_coi_offset(data, offset)
}
\arguments{
\item{data}{List of lists of numeric vectors, where each list
element is a collection of observations across samples at a
single genetic locus.}

\item{offset}{Numeric offset -- n'th highest number of observed alleles}
}
\value{
Numeric vector with one naive COI estimate per sample.
}
\description{
Calculate naive COI offset
}
\details{
Estimates the complexity of infection using a naive approach
that chooses the n'th highest number of observed alleles.
}
\examples{
calculate_naive_coi_offset(simulated_data$data, offset = 2)
}
