% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{combine_simulated_data}
\alias{combine_simulated_data}
\title{Combine simulated data from two populations}
\usage{
combine_simulated_data(simulated_data, simulated_data2)
}
\arguments{
\item{simulated_data}{simulated data from the first population}

\item{simulated_data2}{simulated data from the second population}
}
\value{
Combined simulated data
}
\description{
Combine simulated data from two populations
}
\details{
Combine simulated data from two populations
}
