% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_effective_coi_vectors}
\alias{get_effective_coi_vectors}
\title{Summarize effective COI}
\usage{
get_effective_coi_vectors(chain)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior
distribution to return}

\item{upper_quantile}{The upper quantile of the posterior
distribution to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\description{
Summarize effective COI
}
\details{
Summarize effective COI from MCMC. Returns
a dataframe that contains summaries of the posterior
distribution of effective COI for each biological sample.
}
