% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{get_relatedness_vectors}
\alias{get_relatedness_vectors}
\title{Summarize relatedness}
\usage{
get_relatedness_vectors(chain)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior
distribution to return}

\item{upper_quantile}{The upper quantile of the posterior
distribution to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\description{
Summarize relatedness
}
\details{
Summarize relatedness results from MCMC. Returns
a dataframe that contains summaries of the posterior
distribution of relatedness for each biological sample.
}
