% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{namibia_data}
\alias{namibia_data}
\title{Genetic and epidemiological data from Namibia}
\format{
A data frame with 7 columns and 97214 rows:
\describe{
\item{sample_id}{Sample ID}
\item{HealthFacility}{Health facility}
\item{HealthDistrict}{Health district}
\item{Region}{Region}
\item{Country}{Country}
\item{locus}{Genetic locus}
\item{allele}{Allele observed}
}
}
\source{
\doi{10.7554/eLife.43510.018}
}
\usage{
namibia_data
}
\description{
A dataset containing the genetic and epidemiological data from Namibia
}
\keyword{datasets}
