% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_allele_frequencies}
\alias{regional_allele_frequencies}
\title{Allele frequencies for different regions}
\format{
A list of lists, where each list element is a list of allele frequencies
for a specific region.
}
\usage{
regional_allele_frequencies
}
\description{
A list of allele frequencies for different regions, estimated from the pf7k dataset.
}
\keyword{datasets}
