% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data that is compatible with run_mcmc}
\usage{
sim_data(mean_moi, num_loci, locus_freq_alpha, num_samples, epsilon_pos,
  epsilon_neg)
}
\arguments{
\item{mean_moi}{Mean multiplicity of infection drawn from a poisson}

\item{num_loci}{Number of genetic loci to simulate}

\item{locus_freq_alpha}{List of alphas to be used to simulate from a dirichlet distribution to generate allele frequencies. Generates num_loci distributions for each alpha}

\item{num_samples}{Total number of biological samples to simulate}

\item{epsilon_pos}{False positive rate}

\item{epsilon_neg}{False negative rate}
}
\value{
Simulated data that is structured to go into the MCMC sampler
}
\description{
Simulate data that is compatible with run_mcmc
}
