% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_observed_genotype}
\alias{simulate_observed_genotype}
\title{Simulate observed genotypes}
\usage{
simulate_observed_genotype(
  true_genotypes,
  epsilon_pos,
  epsilon_neg,
  missingness
)
}
\arguments{
\item{true_genotypes}{a list of numeric vectors that are input
to sim_observed_allele}

\item{epsilon_pos}{expected number of false positives}

\item{epsilon_neg}{expected number of false negatives}

\item{missingness}{probability of data being missing}
}
\value{
List the same length as \code{true_genotypes}, each element a binary
vector of observed alleles as returned by \code{\link[=simulate_observed_allele]{simulate_observed_allele()}}.
}
\description{
Simulate observed genotypes
}
\details{
Simulate the observation process across a list of observation
vectors
}
\examples{
true_genotypes <- list(c(1, 0, 1), c(0, 2, 0))
simulate_observed_genotype(
  true_genotypes,
  epsilon_pos = 0.01, epsilon_neg = 0.1, missingness = 0
)
}
