% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_sample_coi}
\alias{simulate_sample_coi}
\title{Simulate sample COI}
\usage{
simulate_sample_coi(num_samples, mean_coi)
}
\arguments{
\item{num_samples}{the total number of biological samples to simulate}

\item{mean_coi}{mean multiplicity of infection}
}
\value{
Integer vector of length \code{num_samples} with a COI of at least 1
for each sample.
}
\description{
Simulate sample COI
}
\details{
Simulate sample COIs from a zero-truncated Poisson distribution
}
\examples{
simulate_sample_coi(num_samples = 10, mean_coi = 2)
}
