% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_sample_moi}
\alias{simulate_sample_moi}
\title{Simulate sample MOIs from a zero-truncated Poisson distribution}
\usage{
simulate_sample_moi(num_samples, mean_moi)
}
\arguments{
\item{num_samples}{the total number of biological samples to simulate}

\item{mean_moi}{mean multiplicity of infection}
}
\description{
Simulate sample MOIs from a zero-truncated Poisson distribution
}
