% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_allele_freq_fn}
\alias{summarize_allele_freq_fn}
\title{Summarize Function of Allele Frequencies}
\usage{
summarize_allele_freq_fn(
  mcmc_results,
  fn,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{fn}{Function that takes as input a simplex to apply to each
allele frequency vector}

\item{lower_quantile}{The lower quantile of the posterior distribution
to return}

\item{upper_quantile}{The upper quantile of the posterior distribution
to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
Data frame with one row per locus (per chain if \code{merge_chains} is
\code{FALSE}) giving posterior quantiles and mean of \code{fn} applied to the sampled
allele frequencies.
}
\description{
Summarize Function of Allele Frequencies
}
\details{
General function to summarize the posterior distribution of
functions of the sampled allele frequencies
}
\examples{
# Posterior summary of the major allele frequency at each locus
major_allele <- summarize_allele_freq_fn(mcmc_results, fn = max)
head(major_allele)
}
