% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_allele_freqs}
\alias{summarize_allele_freqs}
\title{Summarize allele frequencies}
\usage{
summarize_allele_freqs(
  mcmc_results,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior distribution
to return}

\item{upper_quantile}{The upper quantile of the posterior distribution
to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
Data frame with one row per allele (per chain if \code{merge_chains} is
\code{FALSE}) giving posterior quantiles and mean of its frequency, along with
\code{locus} and \code{allele} identifiers.
}
\description{
Summarize allele frequencies
}
\details{
Summarize individual allele frequencies from the posterior
distribution of sampled allele frequencies
}
\examples{
allele_freq_summary <- summarize_allele_freqs(mcmc_results)
head(allele_freq_summary)
}
