% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_coi}
\alias{summarize_coi}
\title{Summarize COI}
\usage{
summarize_coi(
  mcmc_results,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  naive_offset = 2,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc}

\item{lower_quantile}{The lower quantile of the posterior
distribution to return}

\item{upper_quantile}{The upper quantile of the posterior
distribution to return}

\item{naive_offset}{Offset used in calculate_naive_coi_offset}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
Data frame with one row per sample (per chain if \code{merge_chains} is
\code{FALSE}) giving posterior quantiles and mean of COI, naive estimates, and
the posterior probability that the sample is polyclonal.
}
\description{
Summarize COI
}
\details{
Summarize complexity of infection results from MCMC. Returns
a dataframe that contains summaries of the posterior
distribution of COI for each biological sample, as well as naive
estimates of COI.
}
\examples{
coi_summary <- summarize_coi(mcmc_results)
head(coi_summary)
}
