% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_eda}
\alias{summarize_eda}
\title{Summarize expected number of distinct alleles}
\usage{
summarize_eda(
  mcmc_results,
  n,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{n}{values of multiplicity of infection to evaluate the expected number of distinct alleles}

\item{lower_quantile}{The lower quantile of the posterior distribution
to return}

\item{upper_quantile}{The upper quantile of the posterior distribution
to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\description{
Summarize expected number of distinct alleles
}
\details{
Summarize expected number of distinct alleles results from MCMC estimates of allele frequencies.
Returns a dataframe that contains summaries of the posterior distribution of expected number of distinct alleles
}
