% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_effective_coi}
\alias{summarize_effective_coi}
\title{Summarize effective COI}
\usage{
summarize_effective_coi(
  mcmc_results,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior
distribution to return}

\item{upper_quantile}{The upper quantile of the posterior
distribution to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
Data frame with one row per sample (per chain if \code{merge_chains} is
\code{FALSE}) giving posterior quantiles and mean of effective COI,
\code{(1 - relatedness) * (COI - 1) + 1}.
}
\description{
Summarize effective COI
}
\details{
Summarize effective COI from MCMC. Returns
a dataframe that contains summaries of the posterior
distribution of effective COI for each biological sample.
}
\examples{
effective_coi_summary <- summarize_effective_coi(mcmc_results)
head(effective_coi_summary)
}
