% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_epsilon_pos}
\alias{summarize_epsilon_pos}
\title{Summarize epsilon_pos}
\usage{
summarize_epsilon_pos(
  mcmc_results,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior
distribution to return}

\item{upper_quantile}{The upper quantile of the posterior
distribution to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
Data frame with one row per sample (per chain if \code{merge_chains} is
\code{FALSE}) giving posterior quantiles and mean of the false positive rate.
}
\description{
Summarize epsilon_pos
}
\details{
Summarize epsilon positive results from MCMC. Returns
a dataframe that contains summaries of the posterior
distribution of epsilon positive for each biological sample.
}
\examples{
eps_pos_summary <- summarize_epsilon_pos(mcmc_results)
head(eps_pos_summary)
}
