% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_relatedness}
\alias{summarize_relatedness}
\title{Summarize relatedness}
\usage{
summarize_relatedness(
  mcmc_results,
  lower_quantile = 0.025,
  upper_quantile = 0.975,
  merge_chains = TRUE
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior
distribution to return}

\item{upper_quantile}{The upper quantile of the posterior
distribution to return}

\item{merge_chains}{boolean indicating that all chain results should be merged}
}
\value{
Data frame with one row per sample (per chain if \code{merge_chains} is
\code{FALSE}) giving posterior quantiles and mean of within-host relatedness,
computed over draws where COI is greater than 1.
}
\description{
Summarize relatedness
}
\details{
Summarize relatedness results from MCMC. Returns
a dataframe that contains summaries of the posterior
distribution of relatedness for each biological sample.
}
\examples{
relatedness_summary <- summarize_relatedness(mcmc_results)
head(relatedness_summary)
}
