% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_he}
\alias{calculate_he}
\title{Calculate the expected heterozygosity from allele frequencies}
\usage{
calculate_he(allele_freqs)
}
\arguments{
\item{allele_freqs}{Simplex of allele frequencies}
}
\description{
Calculate the expected heterozygosity from allele frequencies
}
