% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{calculate_naive_coi}
\alias{calculate_naive_coi}
\title{Calculate naive COI}
\usage{
calculate_naive_coi(data)
}
\arguments{
\item{data}{List of lists of numeric vectors, where each list
element is a collection of observations across samples at a
single genetic locus.}
}
\description{
Calculate naive COI
}
\details{
Estimates the complexity of infection using a naive approach
that chooses the highest number of observed alleles.
}
