% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_naive_coi_offset}
\alias{calculate_naive_coi_offset}
\title{Estimates the complexity of infection using a naive approach that chooses the n'th highest number of observed alleles.}
\usage{
calculate_naive_coi_offset(data, offset)
}
\arguments{
\item{data}{List of lists of numeric vectors, where each list element is a collection of observations across samples at a single genetic locus.}

\item{offset}{Numeric offset -- n'th highest number of observed alleles}
}
\description{
Estimates the complexity of infection using a naive approach that chooses the n'th highest number of observed alleles.
}
