% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_delimited_data}
\alias{load_delimited_data}
\title{Load delimited data}
\usage{
load_delimited_data(
  data,
  sep = ";",
  warn_uninformative = TRUE,
  aggregate = c("binary", "count")
)
}
\arguments{
\item{data}{data.frame containing the described data}

\item{sep}{string used to separate alleles}

\item{warn_uninformative}{boolean whether or not to print message when
removing uninformative loci}

\item{aggregate}{forwarded to \code{\link[=load_long_form_data]{load_long_form_data()}}; \code{"binary"} (default) or \code{"count"}.}
}
\description{
Load delimited data
}
\details{
Load \code{data.frame} with a \code{sample_id} column and the remaining
columns are \code{loci}. Each cell contains a separator delimited string
representing the observed alleles at that locus for that sample.
Returned data contains vectors \code{sample_ids} and \code{loci} that are ordered
as the results will be ordered from running the MCMC algorithm.
}
