% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_delimited_data}
\alias{load_delimited_data}
\title{Load delimited data}
\usage{
load_delimited_data(data, sep = ";")
}
\arguments{
\item{data}{data.frame containing the described data}

\item{sep}{string used to separate alleles}
}
\description{
Load delimited data
}
\details{
Load \code{data.frame} with a \code{sample_id} column and the remaining
columns are \code{loci}. Each cell contains a separator delimited string
representing the observed alleles at that locus for that sample.
Returned data contains vectors \code{sample_ids} and \code{loci} that are ordered
as the results will be ordered from running the MCMC algorithm.
}
