% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_long_form_data}
\alias{load_long_form_data}
\title{Load long form data}
\usage{
load_long_form_data(df)
}
\arguments{
\item{df}{data frame with 3 columns: \code{sample_id}, \code{locus}, \code{allele}.
Each row is a single observation of an allele at a particular
locus for a given sample.}
}
\description{
Load long form data
}
\details{
Long form data is a data frame with
3 columns: \code{sample_id}, \code{locus}, \code{allele}. Returned data contains
vectors \code{sample_ids} and \code{loci} that are ordered as the results
will be ordered from running the MCMC algorithm.
}
