% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcmc_results}
\alias{mcmc_results}
\title{MCMC results from using the packaged simulated data and calling run_mcmc()}
\format{
An object of class \code{list} of length 8.
}
\usage{
mcmc_results
}
\description{
MCMC results from using the packaged simulated data and calling run_mcmc()
}
\keyword{datasets}
