% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_chain_swaps}
\alias{plot_chain_swaps}
\title{Plot chain swap acceptance rates}
\usage{
plot_chain_swaps(mcmc_results)
}
\arguments{
\item{mcmc_results}{list of results from \code{run_mcmc}}
}
\value{
list of ggplot objects
}
\description{
Plot chain swap acceptance rates
}
\details{
Plot the swap acceptance rates for each chain.
The x-axis is the temperature, and the y-axis is the swap acceptance rate.
The dashed lines indicate the temperatures used for parallel tempering.
}
