% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Implementation of random sampling from a Dirichlet distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{total number of draws}

\item{alpha}{vector controlling the concentration of simplex}
}
\description{
Implementation of random sampling from a Dirichlet distribution
}
