% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Sample from the target distribution using MCMC}
\usage{
run_mcmc(
  data,
  is_missing = FALSE,
  allow_relatedness = TRUE,
  thin = 1,
  burnin = 10000,
  samples_per_chain = 1000,
  verbose = TRUE,
  use_message = FALSE,
  eps_pos_alpha = 1,
  eps_pos_beta = 1,
  eps_neg_alpha = 1,
  eps_neg_beta = 1,
  r_alpha = 1,
  r_beta = 1,
  population_coi_p_alpha = 1,
  population_coi_p_beta = 1,
  population_coi_r_shape = 0.1,
  population_coi_r_rate = 10,
  population_coi_lam_shape = 0.1,
  population_coi_lam_rate = 10,
  population_coi_method = "nb",
  num_populations = 1,
  populations_prior = NULL,
  max_eps_pos = 2,
  max_eps_neg = 2,
  max_coi = 40,
  record_latent_genotypes = FALSE,
  num_chains = 1,
  num_cores = 1,
  pt_chains = 1,
  pt_grad = 1,
  pt_num_threads = 1,
  adapt_temp = TRUE,
  pre_adapt_steps = 25,
  temp_adapt_steps = 25,
  max_initialization_tries = 10000,
  max_runtime = Inf,
  initial_allele_frequencies = NULL,
  observation_model = c("binary", "counts")
)
}
\arguments{
\item{data}{Data to be used in MCMC, as generated by the \verb{load_*_data} functions}

\item{is_missing}{Boolean matrix indicating whether the observation
should be treated as missing data and ignored. Number of rows equals the
number of loci, number of columns equals the number samples. Alternatively,
the user may pass in FALSE if no data should be considered missing.}

\item{allow_relatedness}{Bool indicating whether or not to allow relatedness
within host}

\item{thin}{Positive Integer. How often to sample from mcmc,
1 means do not thin}

\item{burnin}{Positive Integer. Number of MCMC samples to
discard as burnin}

\item{samples_per_chain}{Positive Integer. Number of samples to take
after burnin}

\item{verbose}{Logical indicating if progress is printed}

\item{use_message}{Logical indicating if progress is printed using message or print}

\item{eps_pos_alpha}{Positive Numeric. Alpha parameter in
Beta distribution for eps_pos prior}

\item{eps_pos_beta}{Positive Numeric. Beta parameter in
Beta distribution for eps_pos prior}

\item{eps_neg_alpha}{Positive Numeric. Alpha parameter in
Beta distribution for eps_neg prior}

\item{eps_neg_beta}{Positive Numeric. Beta parameter in
Beta distribution for eps_neg prior}

\item{r_alpha}{Positive Numeric. Alpha parameter in Beta
distribution for relatedness prior}

\item{r_beta}{Positive Numeric. Beta parameter in Beta
distribution for relatedness prior}

\item{population_coi_p_alpha}{Alpha for Beta prior on negative-binomial p (COI)}

\item{population_coi_p_beta}{Beta for Beta prior on negative-binomial p (COI)}

\item{population_coi_r_shape}{Shape for gamma prior on negative-binomial r (COI)}

\item{population_coi_r_rate}{Rate for gamma prior on negative-binomial r (COI)}

\item{population_coi_lam_shape}{Shape for gamma prior on Poisson rate (COI)}

\item{population_coi_lam_rate}{Rate for gamma prior on Poisson rate (COI)}

\item{population_coi_method}{Character. Method for modeling population COI.
For negative binomial model, accepts: "nb", "nbinom", "negbinom", "negative_binomial".
For Poisson model, accepts: "pois", "poisson".}

\item{num_populations}{Number of populations}

\item{populations_prior}{Prior probability for each population (length \code{num_populations} or 1)}

\item{max_eps_pos}{Numeric. Maximum allowed value for eps_pos}

\item{max_eps_neg}{Numeric. Maximum allowed value for eps_neg}

\item{max_coi}{Positive Numeric. Maximum allowed complexity of infection}

\item{record_latent_genotypes}{Logical indicating whether or not to record
the latent genotypes at each step of the MCMC. WARNING: This will increase
the size of the output object significantly.}

\item{num_chains}{Total number of chains to run, possibly simultaneously}

\item{num_cores}{Total OMP parallel threads to use to run chains.
num_cores * pt_num_threads should not exceed the number of cores available
on your system.}

\item{pt_chains}{Total number of chains to run with parallel tempering or a
vector containing the temperatures that should be used for parallel tempering.}

\item{pt_grad}{Power to raise parallel tempering chains to. A value of 1
results in evenly distributed temperatures between [0,1], below 1 will bias
towards 1 and above 1 will bias towards 0. Only used if pt_chains is a
single value (i.e. not a vector).}

\item{pt_num_threads}{Total number of OMP parallel threads to be used to
process parallel tempered chains
num_cores * pt_num_threads should not exceed the number of cores available
on your system.}

\item{adapt_temp}{Logical indicating whether or not to adapt the parallel
tempering temperatures. If TRUE, the temperatures will be adapted during the
\code{burnin} period, starting after \code{pre_adapt_steps} steps. The adaptation will
occur every \code{temp_adapt_steps} steps until burnin is complete. The range of
temperatures will remain the same as specified by \code{pt_chains}.}

\item{pre_adapt_steps}{Number of steps to take before starting to adapt the
parallel tempering temperatures. Only used if \code{adapt_temp} is TRUE.}

\item{temp_adapt_steps}{Number of steps to take between temperature
adaptation steps. Only used if \code{adapt_temp} is TRUE.}

\item{max_initialization_tries}{Number of times to try to initialize the
chain before giving up}

\item{max_runtime}{Maximum runtime in minutes. If the MCMC is running for
more than this amount of time, the function will stop and return the current
state of the MCMC.}

\item{initial_allele_frequencies}{Optional list of initial allele frequencies
for each population. If provided, should be a list of length \code{num_populations},
where each element is a list of length \code{num_loci}, and each locus element is
a numeric vector of allele frequencies that sum to 1. If NULL, allele frequencies
will be initialized using clustering-based approach.}

\item{observation_model}{\code{"binary"} (default) for presence/absence error
process, or \code{"counts"} for Poisson count emissions.}
}
\description{
Sample from the target distribution using MCMC
}
