% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run the MCMC}
\usage{
run_mcmc(data, thin = 0, burnin = 10000, samples = 10000,
  num_chains = 1, importance_sampling_depth = 20, max_coi = 30,
  eps_pos_0 = 0.01, eps_pos_var = 0.25, eps_pos_alpha = 2,
  eps_pos_beta = 100, eps_neg_0 = 0.05, eps_neg_var = 0.25,
  eps_neg_alpha = 6, eps_neg_beta = 96, max_eps_pos = 0.2,
  max_eps_neg = 0.2)
}
\arguments{
\item{data}{Data to be used in MCMC}

\item{thin}{How often to sample from mcmc, 0 means do not thin}

\item{burnin}{Number of MCMC samples to discard as burnin}

\item{samples}{Number of samples to take after burnin}

\item{num_chains}{Number of concurrent chains to run (not implemented)}

\item{importance_sampling_depth}{How many samples to take during importance sampling}

\item{max_coi}{Maximum allowed complexity of infection}

\item{eps_pos_0}{Initial eps_pos value}

\item{eps_pos_var}{Variance used in sampling eps_pos}

\item{eps_pos_alpha}{Alpha parameter in Beta distribution for eps_pos prior}

\item{eps_pos_beta}{Beta parameter in Beta distribution for eps_pos prior}

\item{eps_neg_0}{Initial eps_neg value}

\item{eps_neg_var}{Variance used in sampling eps_neg}

\item{eps_neg_alpha}{Alpha parameter in Beta distribution for eps_neg prior}

\item{eps_neg_beta}{Beta parameter in Beta distribution for eps_neg prior}

\item{max_eps_pos}{Maximum allowed value for eps_pos}

\item{max_eps_neg}{Maximum allowed value for eps_neg}
}
\description{
Run the MCMC
}
