% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Sample from the target distribution using MCMC}
\usage{
run_mcmc(
  data,
  thin = 0,
  burnin = 10000,
  samples = 10000,
  importance_sampling_depth = 100,
  max_coi = 30,
  eps_pos_0 = 0.01,
  eps_pos_var = 0.25,
  eps_pos_alpha = 2,
  eps_pos_beta = 100,
  eps_neg_0 = 0.05,
  eps_neg_var = 0.25,
  eps_neg_alpha = 6,
  eps_neg_beta = 96,
  max_eps_pos = 0.2,
  max_eps_neg = 0.2
)
}
\arguments{
\item{data}{Data to be used in MCMC, expected input is a list of lists of numeric vectors, where each list element is a collection of observations across samples at a single genetic locus.}

\item{thin}{Positive Integer. How often to sample from mcmc, 0 means do not thin}

\item{burnin}{Positive Integer. Number of MCMC samples to discard as burnin}

\item{samples}{Positive Integer. Number of samples to take after burnin}

\item{importance_sampling_depth}{Positive Integer. How many samples to take during importance sampling. Larger values result in longer computation time, too small of values will result in poor mixing.}

\item{max_coi}{Integer. Maximum allowed complexity of infection}

\item{eps_pos_0}{0-1 Numeric. Initial eps_pos value}

\item{eps_pos_var}{0-1 Numeric. Variance used in sampling eps_pos}

\item{eps_pos_alpha}{Positive Numeric. Alpha parameter in Beta distribution for eps_pos prior}

\item{eps_pos_beta}{Positive Numeric. Beta parameter in Beta distribution for eps_pos prior}

\item{eps_neg_0}{0-1 Numeric. Initial eps_neg value}

\item{eps_neg_var}{0-1 Numeric. Variance used in sampling eps_neg}

\item{eps_neg_alpha}{Positive Numeric. Alpha parameter in Beta distribution for eps_neg prior}

\item{eps_neg_beta}{Positive Numeric. Beta parameter in Beta distribution for eps_neg prior}

\item{max_eps_pos}{0-1 Numeric. Maximum allowed value for eps_pos}

\item{max_eps_neg}{0-1 Numeric. Maximum allowed value for eps_neg}
}
\description{
Sample from the target distribution using MCMC
}
