% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Sample from the target distribution using MCMC}
\usage{
run_mcmc(
  data,
  sample_ids,
  loci,
  is_missing = FALSE,
  thin = 1,
  burnin = 10000,
  samples_per_chain = 1000,
  verbose = TRUE,
  allele_freq_threshold = 1e-12,
  eps_pos_0 = 0.1,
  eps_pos_var = 1,
  eps_pos_alpha = 0.1,
  eps_pos_beta = 9.9,
  eps_neg_0 = 0.1,
  eps_neg_var = 1,
  eps_neg_alpha = 0.1,
  eps_neg_beta = 9.9,
  max_eps_pos = 2,
  max_eps_neg = 2,
  max_coi = 20,
  allele_freq_vars = 1,
  adapt_allele_freq_vars = TRUE,
  num_chains = 1,
  num_cores = 1
)
}
\arguments{
\item{data}{Data to be used in MCMC, expected input is a list
of lists of numeric vectors, where each list element is a
collection of observations across samples at a single genetic locus.}

\item{sample_ids}{Vector containing the ordered unique identifier for samples}

\item{loci}{Vector containing the ordered unique identifer for loci}

\item{is_missing}{Boolean matrix indicating whether the observation
should be treated as missing data and ignored. Number of rows equals the
number of loci, number of columns equals the number samples. Alternatively,
the user may pass in FALSE if no data should be considered missing.}

\item{thin}{Positive Integer. How often to sample from mcmc,
1 means do not thin}

\item{burnin}{Positive Integer. Number of MCMC samples to
discard as burnin}

\item{samples_per_chain}{Positive Integer. Number of samples to take
after burnin}

\item{verbose}{Logical indicating if progress is printed}

\item{allele_freq_threshold}{0-1 Numeric. Lowest allowed value for an
allele frequency.}

\item{eps_pos_0}{Numeric. Initial eps_pos value}

\item{eps_pos_var}{Numeric. Variance used in sampling eps_pos}

\item{eps_pos_alpha}{Positive Numeric. Alpha parameter in
Beta distribution for eps_pos prior}

\item{eps_pos_beta}{Positive Numeric. Beta parameter in
Beta distribution for eps_pos prior}

\item{eps_neg_0}{Numeric. Initial eps_neg value}

\item{eps_neg_var}{Numeric. Variance used in sampling eps_neg}

\item{eps_neg_alpha}{Positive Numeric. Alpha parameter in
Beta distribution for eps_neg prior}

\item{eps_neg_beta}{Positive Numeric. Beta parameter in
Beta distribution for eps_neg prior}

\item{max_eps_pos}{Numeric. Maximum allowed value for eps_pos}

\item{max_eps_neg}{Numeric. Maximum allowed value for eps_neg}

\item{max_coi}{Positive Numeric. Maximum allowed complexity of infection}

\item{allele_freq_vars}{Positive Numeric. Variance used in sampling allele
frequencies}

\item{adapt_allele_freq_vars}{Logical indicating whether to adapt variance
to achieve an acceptance rate of .27}

\item{num_chains}{Total number of chains to run, possibly simultaneously}

\item{num_cores}{Total cores to use to run chains}
}
\description{
Sample from the target distribution using MCMC
}
