% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_allele_frequencies}
\alias{simulate_allele_frequencies}
\title{Simulate allele frequencies}
\usage{
simulate_allele_frequencies(alpha, num_loci)
}
\arguments{
\item{alpha}{vector parameter controlling the Dirichlet distribution}

\item{num_loci}{total number of loci to draw}
}
\description{
Simulate allele frequencies
}
\details{
Simulate allele frequency vectors as a draw from a Dirichlet
distribution
}
