% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data generated according to the assumed model}
\usage{
simulate_data(
  mean_coi = NULL,
  num_samples,
  epsilon_pos,
  epsilon_neg,
  sample_cois = NULL,
  locus_freq_alphas = NULL,
  allele_freqs = NULL,
  internal_relatedness_alpha = 0,
  internal_relatedness_beta = 1,
  internal_relatedness = NULL,
  missingness = 0
)
}
\arguments{
\item{mean_coi}{Mean multiplicity of infection drawn from a Poisson}

\item{num_samples}{Total number of biological samples to simulate}

\item{epsilon_pos}{False positive rate, expected number of false positives}

\item{epsilon_neg}{False negative rate, expected number of false negatives}

\item{sample_cois}{List of sample COIs to be used instead of simulating}

\item{locus_freq_alphas}{List of alpha vectors to be used to simulate
from a Dirichlet distribution to generate allele frequencies.}

\item{allele_freqs}{List of allele frequencies to be used instead of
simulating allele frequencies}

\item{internal_relatedness_alpha}{alpha parameter of beta distribution controlling
the random relatedness draws for each sample}

\item{internal_relatedness_beta}{beta parameter of beta distribution controlling
the random relatedness draws for each sample}

\item{internal_relatedness}{List of internal relatedness values to be used
instead of simulating}

\item{missingness}{probability of data being missing}
}
\value{
Simulated data that is structured to go into the MCMC sampler
}
\description{
Simulate data generated according to the assumed model
}
