% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data generated according to the assumed model}
\usage{
simulate_data(
  mean_coi,
  locus_freq_alphas,
  num_samples,
  epsilon_pos,
  epsilon_neg
)
}
\arguments{
\item{mean_coi}{Mean multiplicity of infection drawn from a Poisson}

\item{locus_freq_alphas}{List of alpha vectors to be used to simulate
from a Dirichlet distribution to generate allele frequencies.}

\item{num_samples}{Total number of biological samples to simulate}

\item{epsilon_pos}{False positive rate, between 0 and 1}

\item{epsilon_neg}{False negative rate, between 0 and 1}
}
\value{
Simulated data that is structured to go into the MCMC sampler
}
\description{
Simulate data generated according to the assumed model
}
