% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_observed_genotype}
\alias{simulate_observed_genotype}
\title{Simulate the observation process across a list of observation vectors}
\usage{
simulate_observed_genotype(true_genotypes, epsilon_pos, epsilon_neg)
}
\arguments{
\item{true_genotypes}{a list of numeric vectors that are input to sim_observed_allele}

\item{epsilon_pos}{false positive rate}

\item{epsilon_neg}{false negative rate}
}
\description{
Simulate the observation process across a list of observation vectors
}
