% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulate_observed_genotype}
\alias{simulate_observed_genotype}
\title{Simulate observed genotypes}
\usage{
simulate_observed_genotype(
  true_genotypes,
  epsilon_pos,
  epsilon_neg,
  missingness
)
}
\arguments{
\item{true_genotypes}{a list of numeric vectors that are input
to sim_observed_allele}

\item{epsilon_pos}{expected number of false positives}

\item{epsilon_neg}{expected number of false negatives}

\item{missingness}{probability of data being missing}
}
\description{
Simulate observed genotypes
}
\details{
Simulate the observation process across a list of observation
vectors
}
