% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated genotyping data}
\format{
An object of class \code{list} of length 8.
}
\usage{
simulated_data
}
\description{
A simulated dataset created using \code{simulate_data()}
}
\keyword{datasets}
