% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated genotyping data}
\format{
A list of lists where each list is a collection of genomic data observations at a single locus.
}
\usage{
simulated_data
}
\description{
A simulated dataset created using simulate_data()
}
\keyword{datasets}
