% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_allele_freqs}
\alias{summarize_allele_freqs}
\title{Summarize allele frequencies}
\usage{
summarize_allele_freqs(
  mcmc_results,
  lower_quantile = 0.025,
  upper_quantile = 0.975
)
}
\arguments{
\item{mcmc_results}{Result of calling run_mcmc()}

\item{lower_quantile}{The lower quantile of the posterior distribution
to return}

\item{upper_quantile}{The upper quantile of the posterior distribution
to return}
}
\description{
Summarize allele frequencies
}
\details{
Summarize individual allele frequencies from the posterior
distribution of sampled allele frequencies
}
